Initial state
Added a database with one table called users. users have 4 fields, and their primary key is their user_id which is auto-generated. The three other fields are their username, their password, and their email. As of now, the passwords are in cleartext, I aim to hash these. 
There is also a populate-database python file, which I aim to fill with 5 entries which are bad and always will be bad, and then a number of generated fake-users that will (likely) not be found in the comparison.