Added method to generate 100 users, hardcoding 5 to find as a match.
Removed the hardcoded users as SQL entries. Instead a list of length 100 of type User is created, 5 of these are overwritten with users that should trigger when we then compare them. 
Also added a method to create a random string of variable length, if no argument is sent in it's default 8.