Made a basic web-scraper, added comparison between scraped and local DB.
Added a basic web-scraper tool that gets the list of all 10,000 of Wikipedias most common passwords. Source in the code. Also added a method to get a tuple of all users in the local database, and return the passwords as a set of those. Using the intersection available to sets, a list of passwords found in both list can be found and is indeed 5. 
TODO: 
Fix so there are only 10,000 bad passwords, and not the last list entries on the Wikipedia page with regex. 
Change the regex to only take the passwords of >=8 characters. 
Get a list of the users with bad passwords.